Adds coordinator.applications module
